create user management widget
