+++ war/Manta.html
@@ -76,7 +76,7 @@
     </div>
 
     <div id="authUser">
-      <div id="signUp"></div>
+      <div id="management"></div>
       <div id="Auth"></div>
     </div>
